Removed unused import statements and other minor code smell issues by running intellij's code cleanup function

+++ src/main/java/org/openbooth/gui/specialCells/Double2String.java
@@ -40,6 +40,6 @@ public class Double2String extends StringConverter<Double> {
             return "";
         }
 
-        return (Double.toString(((Double)value).doubleValue()));
+        return (Double.toString(value.doubleValue()));
     }
 }
